+++ config/config.json
@@ -1,8 +1,8 @@
 {
   "development": {
     "username": "root",
-    "password": null,
-    "database": "database_development",
+    "password":"RVLmusic",
+    "database": "dream_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
