+++ routes/dreamers-api-routes.js
@@ -1,4 +1,4 @@
-// var db = require("../models");
+const db = require("../models");
 
 module.exports = function(app) {
   app.get("/api/test", (rep, res) => {
